Cape And Torch Resistance Changes: improve tooltip compatibility with other mods
The current method for generating the tooltip works fine as long as the
mod is the only interacting with `GetDamageModifiersTooltipString`,
however it is not robust in case other mods are also editing tooltips
as they are prone to picking up the Water and Cold custom mods,
resulting in visual issues where the mods appear twice and/or without
label.

To fix this, we filter the Water and Cold custom mods out of the list
passed to the original call, and integrate the missing tooltip
information after the original call has concluded.